Able to upload new live and historic data to the db and request that information via the api

+++ server/src/types.go
@@ -1,8 +1,8 @@
 package main
 
 type ReadingData struct {
-	Time  int64 `json:"time"`
-	Value int   `json:"value"`
+	Time  float64 `json:"time"`
+	Value int     `json:"value"`
 }
 
 type Readings struct {
